rewrite most of components/showObject
