+++ mwdb/web/src/components/ShowObject/common/LatestConfigTab.tsx
@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import { useContext } from "react";
 
 import ConfigTable from "../../ConfigTable";
 
@@ -6,12 +6,20 @@ import { ObjectContext } from "@mwdb-web/commons/context";
 import { ObjectAction, ObjectTab } from "@mwdb-web/commons/ui";
 import { useRemotePath } from "@mwdb-web/commons/remotes";
 
-export default function LatestConfigTab(props) {
+type Props = {
+    label: string;
+};
+
+export function LatestConfigTab(props: Props) {
     const context = useContext(ObjectContext);
     const remotePath = useRemotePath();
 
+    if (!context.object) {
+        return <></>;
+    }
+
     // Don't show tab if object doesn't have latest config
-    if (!context.object.latest_config) return [];
+    if (!context.object.latest_config) return <></>;
     return (
         <ObjectTab
             tab="config"
@@ -27,14 +35,14 @@ export default function LatestConfigTab(props) {
                     )}
                 </div>
             }
-            actions={
+            actions={[
                 <ObjectAction
                     label="Go to config"
                     link={`${remotePath}/config/${context.object.latest_config.id}`}
-                />
-            }
+                />,
+            ]}
             component={() => (
-                <ConfigTable object={context.object.latest_config} />
+                <ConfigTable object={context.object!.latest_config} />
             )}
         />
     );
