+++ mwdb/web/src/types/readable-timestamp.d.ts
@@ -1,0 +1,1 @@
+declare module "readable-timestamp";
